ios,chakra: avoid device memory limits for apps
Adds a limit to the total application memory and checks against it
when allocating more memory in the engine after a threshold.
On iOS, applications have a memory limit after which they are killed,
even if an allocation succeeds. This change should make the Application
more stable during memory spikes, since ChakraCore tries to run the GC
before retrying a failed allocation.